[stats] Adding monthly dialect execution to the stats
Moving the tests to the proper corresponding test file
